todo_text_to_vector/requirements.txt
langchain
langchain-text-splitters
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
torch
torchvision
torchaudio